+++ docs/changelog.md
@@ -20,9 +20,10 @@ Changelog
 =========
 
 *   [Github master](https://github.com/bjones1/CodeChat_Editor):
-    *   No changes.
+    *   Fix indexing in diffs for characters that use more than one UTF-16 code
+        unit, such as 😄,👉🏿,👨‍👦, and 🇺🇳.
 *   v0.1.23, 2025-Jul-24
-    *   Correct diff errors in IDE with CRLF line endings
+    *   Correct diff errors in IDE with CRLF line endings.
     *   Upgrade to newest release of MathJax, TinyMCE.
 *   v0.1.22, 2025-Jul-24 
     *   Better support for opening a page in a web browser.
@@ -123,4 +124,4 @@ Changelog
     *   Cross-platform fixes.
 *   v0.1.0, 2024-Oct-16:
     *   Initial release, with binaries for Windows only. Built with
-        manually-patched CodeMirror per [this
+        manually-patched CodeMirror per \[this