+++ docs/changelog.md
@@ -22,6 +22,7 @@ Changelog
 *   [Github master](https://github.com/bjones1/CodeChat_Editor):
     *   Fix indexing in diffs for characters that use more than one UTF-16 code
         unit, such as 😄,👉🏿,👨‍👦, and 🇺🇳.
+    *   Fix data corruption with adjacent doc blocks.
 *   v0.1.23, 2025-Jul-24
     *   Correct diff errors in IDE with CRLF line endings.
     *   Upgrade to newest release of MathJax, TinyMCE.
@@ -124,4 +125,5 @@ Changelog
     *   Cross-platform fixes.
 *   v0.1.0, 2024-Oct-16:
     *   Initial release, with binaries for Windows only. Built with
-        manually-patched CodeMirror per \[this
+        manually-patched CodeMirror per [this
+        issue](https://github.com/bjones1/CodeChat_Editor/issues/27).